#1741 : time for retry modified + log management
